optimize searching and indexing
merge searcher and indexer into model.rs
