+++ README.md
@@ -15,6 +15,8 @@
 | `idf`       | Inverse Document Frequency          |                                                                                                                                                    |
 | `dtf`       | Document Terms Frequencies          | The HashMap that stores the frequencies of terms that occur in the document, with the term itself as the key and the frequency count as the value. |
 | `dtf_index` | Index of Document Terms Frequencies | The HashMap that stores Document Terms Frequencies with the file path as key.                                                                      |
+| `tcf`       | Term Corpus-wide Frequency          | The number of documents where a term occurs                                                                                                        |
+| `dtc`       | Document Terms Count                | The number of how many terms exist in a document                                                                                                   |
 
 ## TODO
