+++ src/io/writer.rs
@@ -4,12 +4,9 @@ use std::{
     path::Path,
 };
 
-use crate::DocumentTermsFrequenciesIndex;
+use crate::Model;
 
-pub fn save_index(
-    index: &DocumentTermsFrequenciesIndex,
-    output_path: impl AsRef<Path>,
-) -> io::Result<()> {
-    serde_json::to_writer(BufWriter::new(File::create(output_path)?), &index)?;
+pub fn write_model(model: &Model, output_path: impl AsRef<Path>) -> io::Result<()> {
+    serde_json::to_writer(BufWriter::new(File::create(output_path)?), model)?;
     Ok(())
 }
