+++ src/lib.rs
@@ -1,13 +1,5 @@
-mod indexer;
-mod io;
 mod lexer;
 mod model;
-mod searcher;
 
-pub use indexer::{index_dir, index_text};
-
-pub use model::{DocumentTermsFrequencies, DocumentTermsFrequenciesIndex};
-
-pub use searcher::search;
-
-pub use io::{read_index, save_index};
+pub mod io;
+pub use model::Model;
